- Added a Snap To Angle Feature on the Polyline Tool - Resolved bug - Graphical issue when equipping polytool and zooming #1890 - Fixed All build Errors and typos.

+++ core_lib/src/managers/toolmanager.h
@@ -85,7 +85,7 @@ public slots:
     void setCameraPathDotColor(int);
     void resetCameraTransform(CameraFieldOption option);
     void setSnapAngleBox(bool);
-    void setSnapAngleDegrees(double);
+    void setSnapAngleDegrees(int);
 
     /// Layer mode will be enforced by the the choice the reference mode selected.
     /// @return Returns true if reference mode is ``current layer`, otherwise false.
